+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@hookform/resolvers": "^4.1.0",
     "axios": "^1.7.9",
     "js-cookie": "^3.0.5",
     "next": "^15.1.6",
@@ -10,7 +11,8 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
-    "styled-components": "^6.1.14"
+    "styled-components": "^6.1.14",
+    "yup": "^1.6.1"
   },
   "scripts": {
     "dev": "next dev",
@@ -37,6 +39,7 @@
     ]
   },
   "devDependencies": {
+    "@hookform/devtools": "^4.3.3",
     "@types/node": "22.12.0",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
